admin: update - Configure the Vue.js project
The 3rd party libraries 'jQuery', 'bootstrap' and 'Popper.js' will no
longer be loaded by the index.html.
From now on the files get imported by the App.vue.

Adds documentation.

Adds a .htaccess file as an example as well as settings for the dev
server and a new publicPath, when using the app in production.

Adds the Skydivemanifest theme. That means that this commit overrides
the default bootstrap variables.

Adds a login component, also defines the basic routes and sets up a
navigation guard.

Replaces the default app logo and the favicons.

Updates the page title on every route change dynamically or uses the
default if no 'meta.title' is assigned to the route.

Adds i18n functionality. Adds English and German translations in the
same step and embeds them in the Login Component.

Adds temporary fix for eslint.
As far as I can tell babel-eslint has some issues currently. The linting
of template literals is failing. As long as this hasn't been fixed, I
just keep the additional eslint settings.

Defines basic pwa settings and ignores the .htaccess file when building.

Configures additional ESLint rules, enables semicolons.

Configures the test environment.

Updates the 'jest.config.js' to ignore the main.js and
registerServiceWorker.js.
Adds tests for i18n.

Adds grid support for ie 10 and 11.

Adds different layout components that can be set for each route by
adding the meta.layout field in the router.

Signed-off-by: Sebastian Krah <exotelis@mailbox.org>